Start of robotic welding arm
I have a "3D printer" (moving wall of welders with blueprint projector) that misses some pieces. I could create filler blocks to fix it, but I'd like to try filling in the missing pieces with a robotic welding arm. Currently I'm setting it up so that The script gets called repeatedly and it runs through a list of lists of rotations: it waits for each of the parts of the arm to be rotated into place and then runs the next set of rotations.

Is this the best way to do it? Oh heck no. Is it a fun little problem and will it look cool? Yes. I want it to look like a welding robot in a factory. For now this will be specific to one blueprint that I have. I may be able to create a different program for each blueprint that's stored on the ship itself and perhaps fires when the ship is complete, but for now I just want the arm to move into a bunch of different positions.

I'm using the MDK from https://github.com/malware-dev/MDK-SE to get the syntax checking and completion in Visual Studio.

I haven't done C# in well over a decade.

The script is currently in flux and doesn't work. It did work and I started changing it but I want to get it committed so I can start tracking changes with commit comments so I don't lose the history.