All kinds of debugging
I've updated the logging so I can output progress indicators without a newline. This helped me see that I wasn't updated the count unless the count was zero or divisible by ten.

Added the ShouldStop function to see if the rotations list has been executed completely and the script hasn't been run manually: i.e. if the script was run by an UpdateFrequence.Update100 then we don't want to keep running. If ShouldStop returns true then I set the UpdateFrequence to UpdateNone. Once the list of rotations is done then I set "isFinished" to true.

One of the hinges wasn't updating because it was outside of its limits; the game said it wouldn't move until it was back inside the limits. I updated the code to either the upper or lower limit to the current degrees and then the other limit to the target. For some reason the bottom hinge still won't move at all.

Added a ClearLog function so I could blank the screen. Now that I've added shouldNewLine to the Log function I can simplify that down to simply calling Log("", false, false). The function was originally because Log automatically added a "\n".