Move hinges (+90 to -90) differently than rotors
Rotors go all the way around but hinges only go from +90 to -90 degrees.

I'm trying to setup rotors so that they always turn the shortest distance to get to the target degrees.

Something's screwy because the arms aren't going to the start position. I've messed with the torque and the velocity. I suspect that maybe I'm setting the max and min angles wrong so that they rotors can't move.